move Seat.cs and Status.cs into Models/Seats folder, same with SeatTests.cs
