+++ CinnamonCinemas/Models/Theatre.cs
@@ -1,4 +1,5 @@
 ﻿using CinnamonCinemas.Models.SeatNumberGenerators;
+using CinnamonCinemas.Models.Seats;
 using System.Collections.ObjectModel;
 
 namespace CinnamonCinemas.Models;
